# api/openapi.yaml
...
components:
  schemas:
    ErrorResponse:
      properties:
        error:
          type: string
      type: object
...
    StatsResponse:
      properties:
        completed_count:
          type: integer
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        session_count:
          type: integer
        tool_calls:
          type: integer
        total_cost:
          type: number
        total_duration_ms:
          type: integer
        turn_count:
          type: integer
      type: object
...
info:
  contact: {}
  description: |-
    HTTP API for inspecting, querying, and searching stored Tapes sessions.

    The REST surface exposes health checks, session listing and retrieval, derived session summaries, aggregate stats, semantic search, skill authoring and publishing, operator maintenance endpoints, and a streamable MCP endpoint.

    Not covered here: the ingest write surface, which is a separate server with its own contract (see ingest/openapi.yaml).
  title: Tapes API
  version: "1.0"
openapi: 3.0.3
paths:
...
  /v1/stats:
    get:
      description: Returns counts plus cost / token / duration / tool-call / completed-count
        totals for the window. The numbers are span-grain trace rollup sums (delta-only
        usage, agent time = sum of trace durations) so they agree with the session
        and trace views; turn_count counts traces. Filter the window with since/until.
      operationId: getStats
      parameters:
      - description: Only include records at or after this RFC3339 timestamp
        in: query
        name: since
        schema:
          format: date-time
          type: string
      - description: Only include records before or at this RFC3339 timestamp
        in: query
        name: until
        schema:
          format: date-time
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Invalid query parameters
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Failed to compute stats
      summary: Get aggregate session stats
      tags:
      - sessions
...